chore: organize tests
